requirements.txt
torch
numpy
matplotlib
tensorboard
scikit-learn
pandas
torchvision
pillow